architecture-1-serverless-lambda/src/lambda/extract/requirements.txt
boto3==1.34.0
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.1.4
